refactor: reorganize functions and remove index.js
