+++ packages/telemetry/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@webiny/telemetry",
   "version": "5.14.0",
-  "main": "index.js",
   "license": "MIT",
   "dependencies": {
     "@webiny/global-config": "^5.13.0",
