+++ packages/telemetry/sendEvent.js
@@ -1,11 +1,19 @@
 const FormData = require("form-data");
 const fetch = require("node-fetch");
-const { globalConfig } = require("@webiny/global-config");
 
 const API_KEY = "ZdDZgkeOt4Z_m-UWmqFsE1d6-kcCK3BH0ypYTUIFty4";
 const API_URL = "https://t.webiny.com";
 
-module.exports.sendEvent = async ({ event, user, version, properties, extraPayload } = {}) => {
+/**
+ * The main sendEvent function
+ * @param event
+ * @param user
+ * @param version
+ * @param properties
+ * @param extraPayload
+ * @returns {Promise<T>}
+ */
+module.exports = async ({ event, user, version, properties, extraPayload } = {}) => {
     if (!event) {
         throw new Error(`Cannot send event - missing "event" name.`);
     }
@@ -49,15 +57,3 @@ module.exports.sendEvent = async ({ event, user, version, properties, extraPaylo
     });
 };
 
-module.exports.enable = () => {
-    globalConfig.set("telemetry", true);
-};
-
-module.exports.disable = () => {
-    globalConfig.set("telemetry", false);
-};
-
-module.exports.isEnabled = () => {
-    const config = globalConfig.get();
-    return config.telemetry !== false || config.tracking !== false;
-};
